+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 - [#2095](https://github.com/influxdata/kapacitor/issues/2095): Add barrier node support to join node.
+- [#1157](https://github.com/influxdata/kapacitor/issues/1157): Add ability to expire groups using the barrier node.
 
 ### Bugfixes