+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - [#2099](https://github.com/influxdata/kapacitor/issues/2099): Add `alert/persist-topics` to config
 - [#2101](https://github.com/influxdata/kapacitor/issues/2101): Add multiple field support to the change detect node.
 - [#1961](https://github.com/influxdata/kapacitor/pull/1961): Add links to pagerduty2 alerts
+- [#1974](https://github.com/influxdata/kapacitor/issues/1974): Add additional metadata to Sensu alerts.
 
 ### Bugfixes