+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## unreleased
 
+## v1.5.2 [2018-12-12]
+
 ### Features
 
 - [#2095](https://github.com/influxdata/kapacitor/issues/2095): Add barrier node support to join node.